+++ packages/devtools/src/runtime/plugins/view/client.ts
@@ -1,7 +1,223 @@
 import type { Ref } from 'vue'
-import { computed, ref } from 'vue'
+import { computed, createApp, h, markRaw, ref, shallowReactive, shallowRef, watch } from 'vue'
+import { createHooks } from 'hookable'
+import type { NuxtDevtoolsHostClient } from '../../../types'
+import Main from './Main.vue'
+import { closePanel, popupWindow, state, togglePanel } from './state'
+
+// eslint-disable-next-line @typescript-eslint/prefer-ts-expect-error
+// @ts-ignore tsconfig
+import { useAppConfig } from '#imports'
+
+const clientRef = shallowRef<NuxtDevtoolsHostClient>()
+
+export { clientRef as client }
 
 export type ColorScheme = 'dark' | 'light'
+
+export async function setupDevToolsClient({
+  nuxt,
+  clientHooks,
+  timeMetric,
+}: {
+  nuxt: any
+  clientHooks: any
+  timeMetric: any
+}) {
+  const isInspecting = ref(false)
+  const colorMode = useClientColorMode()
+
+  const client: NuxtDevtoolsHostClient = shallowReactive({
+    nuxt: markRaw(nuxt as any),
+    appConfig: useAppConfig() as any,
+    hooks: createHooks(),
+    getClientHooksMetrics: () => Object.values(clientHooks),
+    getClientPluginMetrics: () => {
+      return window.__NUXT_DEVTOOLS_PLUGINS_METRIC__ || []
+    },
+    loadingTimeMetrics: timeMetric,
+    reloadPage() {
+      location.reload()
+    },
+    closeDevTools: closePanel,
+    inspector: getInspectorInstance(),
+    colorMode,
+    getIframe,
+    updateClient,
+  })
+
+  let iframe: HTMLIFrameElement | undefined
+
+  function updateClient() {
+    if (!client.inspector)
+      client.inspector = getInspectorInstance()
+
+    try {
+      iframe?.contentWindow?.__NUXT_DEVTOOLS_VIEW__?.setClient(client)
+    }
+    catch (e) {
+      // cross-origin
+    }
+    return client
+  }
+
+  function getIframe() {
+    if (!iframe) {
+      const CLIENT_PATH = '/__nuxt_devtools__/client'
+      const initialUrl = CLIENT_PATH + state.value.route
+      iframe = document.createElement('iframe')
+      iframe.id = 'nuxt-devtools-iframe'
+      iframe.src = initialUrl
+      iframe.setAttribute('data-v-inspector-ignore', 'true')
+      iframe.addEventListener('load', async () => {
+        await waitForClientInjection()
+        client.updateClient()
+      })
+    }
+
+    return iframe
+  }
+
+  function waitForClientInjection(retry = 10, timeout = 200) {
+    const test = () => !!iframe?.contentWindow?.__NUXT_DEVTOOLS_VIEW__
+
+    if (test())
+      return
+
+    return new Promise<void>((resolve, reject) => {
+      const interval = setInterval(() => {
+        if (test()) {
+          clearInterval(interval)
+          resolve()
+        }
+        else if (retry-- <= 0) {
+          clearInterval(interval)
+          // eslint-disable-next-line prefer-promise-reject-errors
+          reject('Nuxt Devtools client injection failed')
+        }
+      }, timeout)
+    })
+  }
+
+  function enableComponentInspector() {
+    window.__VUE_INSPECTOR__?.enable()
+    isInspecting.value = true
+  }
+
+  function disableComponentInspector() {
+    if (!window.__VUE_INSPECTOR__?.enabled)
+      return
+
+    window.__VUE_INSPECTOR__?.disable()
+    client?.hooks.callHook('host:inspector:close')
+    isInspecting.value = false
+  }
+
+  function getInspectorInstance(): NuxtDevtoolsHostClient['inspector'] {
+    const componentInspector = window.__VUE_INSPECTOR__
+    if (componentInspector) {
+      componentInspector.openInEditor = async (baseUrl, file, line, column) => {
+        disableComponentInspector()
+        await client.hooks.callHook('host:inspector:click', baseUrl, file, line, column)
+      }
+      componentInspector.onUpdated = () => {
+        client.hooks.callHook('host:inspector:update', {
+          ...componentInspector.linkParams,
+          ...componentInspector.position,
+        })
+      }
+    }
+    return markRaw({
+      isEnabled: isInspecting,
+      enable: enableComponentInspector,
+      disable: disableComponentInspector,
+      toggle: () => {
+        if (window.__VUE_INSPECTOR__?.enabled)
+          disableComponentInspector()
+        else
+          enableComponentInspector()
+      },
+      instance: componentInspector,
+    })
+  }
+
+  function refreshReactivity() {
+    client.hooks.callHook('host:update:reactivity')
+  }
+
+  // trigger update for reactivity
+  watch(() => [
+    client.nuxt.payload,
+    client.colorMode.value,
+    client.loadingTimeMetrics,
+  ], () => {
+    refreshReactivity()
+  }, { deep: true })
+  // trigger update for route change
+  client.nuxt.vueApp.config.globalProperties?.$router?.afterEach(() => {
+    refreshReactivity()
+  })
+  // trigger update for app mounted
+  client.nuxt.hook('app:mounted', () => {
+    refreshReactivity()
+  })
+  // record last route
+  client.hooks.hook('devtools:navigate', (path) => {
+    state.value.route = path
+  })
+
+  clientRef.value = client
+
+  // Experimental: Picture-in-Picture mode
+  // https://developer.chrome.com/docs/web-platform/document-picture-in-picture/
+  const documentPictureInPicture = window.documentPictureInPicture
+  if (documentPictureInPicture?.requestWindow) {
+    client.popup = async () => {
+      const iframe = getIframe()
+      const pip = popupWindow.value = await documentPictureInPicture.requestWindow({
+        width: iframe!.clientWidth || 500,
+        height: iframe!.clientHeight || 500,
+      }) as Window
+      const style = pip.document.createElement('style')
+      style.innerHTML = `
+          body {
+            margin: 0;
+            padding: 0;
+          }
+          iframe {
+            width: 100vw;
+            height: 100vh;
+            border: none;
+            outline: none;
+          }
+        `
+      pip.__NUXT_DEVTOOLS_DISABLE__ = true
+      pip.document.head.appendChild(style)
+      pip.document.body.appendChild(iframe)
+    }
+  }
+
+  client.updateClient()
+
+  const holder = document.createElement('div')
+  holder.id = 'nuxt-devtools-container'
+  holder.setAttribute('data-v-inspector-ignore', 'true')
+  document.body.appendChild(holder)
+
+  // Shortcut to toggle devtools
+  addEventListener('keydown', (e) => {
+    if (e.code === 'KeyD' && e.altKey && e.shiftKey)
+      togglePanel()
+  })
+
+  const app = createApp({
+    render: () => h(Main, { client }),
+    devtools: {
+      hide: true,
+    },
+  })
+  app.mount(holder)
+}
 
 export function useClientColorMode(): Ref<ColorScheme> {
   const explicitColor = ref<ColorScheme>()
